feat(view): スケールするコンポーネントを切り出し
